feat(added_elements): этап 2+4 — transform-SQL и backfill-скрипт
Параметризованные SQL для нового ELT-конвейера + одноразовый backfill.

transform/01_extract_load_raw.sql: DELETE+INSERT окна [last, run) в raw_added_elements.
  - revit_ext.added_element  -> source='added',  date >= 2026-02-14
  - revit_ext.modified_element -> source='modified', date >= 2026-02-14
  - legacy_ext.added_element_legacy -> source='legacy', date < 2026-02-14
  Граница 2026-02-14 захардкожена в WHERE через GREATEST/LEAST.

transform/02_transform_staging.sql: обогащение raw -> stg.
  - окно [last - 1 day, run) для корректного LAG на границе сессий
  - LEFT JOIN dim_ad_users / dim_transactions / dim_bim_users
  - LATERAL classify_transaction_fallback для fallback-классификации
  - LAG(date) OVER (PARTITION BY user_name ORDER BY date, id)
  - ROW_NUMBER OVER (user_name, date) для флага дубль-timestamp
  - Полное соответствие логике сессий в Python:
    SESSION_GAP_SECONDS=900, MIN_TRANSACTION_SEC=30,
    обнуление пауз, защита от дубль-строк.

transform/03_build_marts.sql: stg -> ext_added_elements_{bim,designers}.
  - окно [last, run) (БЕЗ буферного дня — он только для STG)
  - DELETE+INSERT обеих витрин в одной транзакции.

scripts/backfill_added_elements.py: автоматический бэкфилл по окнам в 14 дней.
  - Определяет диапазон через FDW (LEAST/GREATEST min/max по 3 источникам).
  - При ошибке логирует команду перезапуска (--start YYYY-MM-DD).
  - Поддерживает --dry-run / --start / --end / --days.

Поправки DDL под прод-совместимость (raw/stg/marts были пустыми):
  - program_version       INTEGER -> BIGINT
  - time_since_prev_sec   INTEGER -> DOUBLE PRECISION
  (Python diff().total_seconds() = float64; INTEGER давал бы расхождение
   на дробных частях секунд с текущими витринами.)

+++ dags/etl_pipelines/sql/added_elements/ddl/01_raw.sql
@@ -35,7 +35,9 @@ CREATE TABLE IF NOT EXISTS datalake.raw_added_elements (
 
     -- Поля, существующие только в revit.added_element / revit.modified_element.
     -- Для source='legacy' остаются NULL.
-    program_version     INTEGER,
+    -- program_version — BIGINT, потому что в текущем Python используется nullable
+    -- Int64 (см. transform/added_elements.py:460), и pandas COPY кладёт это в BIGINT.
+    program_version     BIGINT,
     project_path        TEXT,
     group_model         TEXT,
     description         TEXT,
